module4-task2: * move get-unique-offer-cities helper to global helpers
