+++ src/components/favorite-cities-item/favorite-cities-item.jsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import {offerType} from '~/common/prop-types/prop-types';
+import {offerType, offerCityType} from '~/common/prop-types/prop-types';
 import FavoriteOffersList from '~/components/favorite-offers-list/favorite-offers-list';
 
 const FavoriteCitiesItem = ({city, offers}) => (
@@ -8,7 +8,7 @@ const FavoriteCitiesItem = ({city, offers}) => (
     <div className="favorites__locations locations locations--current">
       <div className="locations__item">
         <span className="locations__item-link">
-          <span>{city}</span>
+          <span>{city.name}</span>
         </span>
       </div>
     </div>
@@ -17,7 +17,7 @@ const FavoriteCitiesItem = ({city, offers}) => (
 );
 
 FavoriteCitiesItem.propTypes = {
-  city: PropTypes.string.isRequired,
+  city: offerCityType.isRequired,
   offers: PropTypes.arrayOf(offerType.isRequired).isRequired,
 };
