+++ src/components/favorite-cities-list/favorite-cities-list.jsx
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import {getUniqueOfferCities} from '~/helpers/helpers';
 import {offerType} from '~/common/prop-types/prop-types';
 import FavoriteCitiesItem from '~/components/favorite-cities-item/favorite-cities-item';
-import {getUniquesCities, getOffersByCity} from './helpers';
+import {getOffersByCity} from './helpers';
 
 const FavoriteCitiesList = ({offers}) => {
-  const uniqueCities = getUniquesCities(offers);
+  const uniqueCities = getUniqueOfferCities(offers);
 
   return (
     <ul className="favorites__list">
