+++ src/components/favorite-cities-list/helpers.js
@@ -1,16 +1,7 @@
-import {getUniqueItems} from '~/helpers/helpers';
-
-const getUniquesCities = (offers) => {
-  const offersCities = offers.map((offer) => offer.city.name);
-  const uniqueCities = getUniqueItems(offersCities);
-
-  return uniqueCities;
-};
-
 const getOffersByCity = (offers, city) => {
-  const offersByCity = offers.filter((offer) => offer.city.name === city);
+  const offersByCity = offers.filter((offer) => offer.city.name === city.name);
 
   return offersByCity;
 };
 
-export {getUniquesCities, getOffersByCity};
+export {getOffersByCity};
